+++ src/main/java/ru/egarcourses/HospitalInfoSystem/services/DoctorService.java
@@ -1,7 +1,6 @@
 package ru.egarcourses.HospitalInfoSystem.services;
 
 import ru.egarcourses.HospitalInfoSystem.dto.DoctorDTO;
-import ru.egarcourses.HospitalInfoSystem.models.Doctor;
 
 import java.util.List;
 
@@ -9,11 +8,11 @@ public interface DoctorService {
 
     public List<DoctorDTO> findAll();
 
-    public DoctorDTO findById(int id);
+    public DoctorDTO findById(Long id);
 
     public void save(DoctorDTO doctorDTO);
 
-    public void update(int id, DoctorDTO updatedDoctorDTO);
+    public void update(Long id, DoctorDTO updatedDoctorDTO);
 
-    public void delete(int id);
+    public void delete(Long id);
 }
